feat: add typed marketplace models and mock data
