+++ src/components/marketplace/trendingExpert.tsx
@@ -1,71 +1,8 @@
 "use client";
 
 import React, { useState } from 'react';
-import { ExternalLink, ChevronLeft, ChevronRight } from 'lucide-react';
+import { mockExpert } from '../../../utils/data/mock-data';
 
-interface ExpertCollection {
-  id: string;
-  title: string;
-  creator: {
-    name: string;
-    avatar: string;
-  };
-  description: string;
-  image: string;
-  price: string;
-  highestBid: string;
-}
-
-const mockData: ExpertCollection[] = [
-  {
-    id: '1',
-    title: 'AstroFiction',
-    creator: {
-      name: 'Spaceone',
-      avatar: '/assets/Avatar.svg',
-    },
-    description: 'Lorem ipsum neque molestie nunc feugiat tincidunt vitae risus suscipit dictum tortor nunc sed.',
-    image: '/assets/space_girl.svg',
-    price: '1.63 ETH',
-    highestBid: '0.33 wETH',
-  },
-  {
-    id: '2',
-    title: 'Science Lab',
-    creator: {
-      name: 'Spaceone',
-      avatar: '/assets/Avatar.svg',
-    },
-    description: 'Lorem ipsum neque molestie nunc feugiat tincidunt vitae risus suscipit dictum tortor nunc sed.',
-    image: '/assets/science_lab_collection.png',
-    price: '2.45 ETH',
-    highestBid: '0.50 wETH',
-  },
-  {
-    id: '3',
-    title: 'Digital Future',
-    creator: {
-      name: 'NeoTech',
-      avatar: '/assets/Avatar.svg',
-    },
-    description: 'Exploring the boundaries of digital reality and blockchain technology in the modern era.',
-    image: '/assets/astronaut_collection.png', // Reusing for mock
-    price: '0.85 ETH',
-    highestBid: '0.12 wETH',
-  },
-  {
-    id: '4',
-    title: 'Quantum Art',
-    creator: {
-      name: 'Visionary',
-      avatar: '/assets/Avatar.svg',
-    },
-    description: 'A collection focusing on the intersection of quantum physics and visual aesthetics.',
-    image: '/assets/science_lab_collection.png', // Reusing for mock
-    price: '3.10 ETH',
-    highestBid: '0.75 wETH',
-  }
-];
 
 export default function TrendingExpert() {
   const [activeSlide, setActiveSlide] = useState(0);
@@ -81,54 +18,53 @@ export default function TrendingExpert() {
 
         {/* Desktop View: Horizontal Scroll with Fade */}
         <div className="hidden md:block relative group">
-          <div 
+          <div
             className="flex space-x-6 overflow-x-auto no-scrollbar scroll-smooth pb-8"
-            style={{ 
-              scrollbarWidth: 'none', 
+            style={{
+              scrollbarWidth: 'none',
               msOverflowStyle: 'none',
               maskImage: 'linear-gradient(to right, black 85%, transparent 100%)',
               WebkitMaskImage: 'linear-gradient(to right, black 85%, transparent 100%)'
             }}
           >
-            {mockData.map((collection, index) => (
-              <div 
+            {mockExpert.map((collection, index) => (
+              <div
                 key={collection.id}
-                className={`flex-shrink-0 relative rounded-3xl overflow-hidden group/card transition-all duration-300 ${
-                  index === 0 ? 'w-[600px] h-[450px]' : 'w-[400px] h-[450px]'
-                }`}
+                className={`flex-shrink-0 relative rounded-3xl overflow-hidden group/card transition-all duration-300 ${index === 0 ? 'w-[600px] h-[450px]' : 'w-[400px] h-[450px]'
+                  }`}
               >
                 {/* Background Image */}
-                <img 
-                  src={collection.image} 
+                <img
+                  src={collection.image}
                   alt={collection.title}
                   className="w-full h-full object-cover transition-transform duration-500 group-hover/card:scale-110"
                 />
-                
+
                 {/* Gradient Overlay */}
                 <div className="absolute inset-0 bg-gradient-to-t from-black/90 via-black/20 to-transparent" />
-                
+
                 {/* Card Content */}
                 <div className="absolute inset-0 p-8 flex flex-col justify-end">
                   <h3 className="text-4xl font-bold mb-4 text-white">{collection.title}</h3>
-                  
+
                   <div className="flex items-center space-x-3 mb-4">
-                    <img 
-                      src={collection.creator.avatar} 
+                    <img
+                      src={collection.creator.avatar}
                       alt={collection.creator.name}
                       className="w-8 h-8 rounded-full border border-gray-600"
                     />
                     <span className="text-sm font-medium text-white">{collection.creator.name}</span>
                   </div>
-                  
+
                   <p className="text-gray-300 text-sm mb-6 line-clamp-2 max-w-sm">
                     {collection.description}
                   </p>
-                  
+
                   <div className="flex items-center justify-between">
                     <button className="px-6 py-2.5 text-white bg-white/10 backdrop-blur-md border border-white/20 rounded-xl text-sm font-medium hover:bg-white/20 transition-colors">
                       Buy now
                     </button>
-                    
+
                     <div className="flex space-x-6">
                       <div className="flex flex-col">
                         <span className="text-[10px] text-gray-500 uppercase tracking-wider">Price</span>
@@ -144,7 +80,7 @@ export default function TrendingExpert() {
               </div>
             ))}
           </div>
-          
+
           {/* Scroll Fade Overlay (Visual only, pointer events none) */}
           <div className="absolute top-0 right-0 h-full w-24 bg-gradient-to-l from-black to-transparent pointer-events-none" />
         </div>
@@ -152,20 +88,20 @@ export default function TrendingExpert() {
         {/* Mobile View: Swipable Slider with Pagination */}
         <div className="md:hidden">
           <div className="relative">
-            <div 
+            <div
               className="flex transition-transform duration-500 ease-out"
               style={{ transform: `translateX(-${activeSlide * 100}%)` }}
             >
-              {mockData.map((collection) => (
+              {mockExpert.map((collection) => (
                 <div key={collection.id} className="w-full flex-shrink-0 px-2">
                   <div className="relative rounded-3xl overflow-hidden h-[500px]">
-                    <img 
-                      src={collection.image} 
+                    <img
+                      src={collection.image}
                       alt={collection.title}
                       className="w-full h-full object-cover"
                     />
                     <div className="absolute inset-0 bg-gradient-to-t from-black/90 via-black/40 to-transparent" />
-                    
+
                     <div className="absolute inset-0 p-6 flex flex-col justify-end">
                       <h3 className="text-3xl font-bold mb-4">{collection.title}</h3>
                       <div className="flex items-center space-x-3 mb-4">
@@ -176,15 +112,15 @@ export default function TrendingExpert() {
                         {collection.description}
                       </p>
                       <div className="flex flex-col gap-4">
-                         <div className="flex items-center justify-between">
-                            <div className="flex flex-col">
-                                <span className="text-[10px] text-gray-500 uppercase font-mono">Price</span>
-                                <span className="text-sm font-bold">{collection.price}</span>
-                            </div>
-                            <div className="flex flex-col text-right">
-                                <span className="text-[10px] text-gray-500 uppercase font-mono">Highest Bid</span>
-                                <span className="text-sm font-bold">{collection.highestBid}</span>
-                            </div>
+                        <div className="flex items-center justify-between">
+                          <div className="flex flex-col">
+                            <span className="text-[10px] text-gray-500 uppercase font-mono">Price</span>
+                            <span className="text-sm font-bold">{collection.price}</span>
+                          </div>
+                          <div className="flex flex-col text-right">
+                            <span className="text-[10px] text-gray-500 uppercase font-mono">Highest Bid</span>
+                            <span className="text-sm font-bold">{collection.highestBid}</span>
+                          </div>
                         </div>
                         <button className="w-full py-3 bg-white/10 backdrop-blur-md border border-white/20 rounded-xl text-sm font-bold">
                           Buy now
@@ -198,13 +134,12 @@ export default function TrendingExpert() {
 
             {/* Pagination Dots */}
             <div className="flex justify-center mt-6 space-x-2">
-              {mockData.map((_, index) => (
+              {mockExpert.map((_, index) => (
                 <button
                   key={index}
                   onClick={() => setActiveSlide(index)}
-                  className={`h-1.5 rounded-full transition-all duration-300 ${
-                    activeSlide === index ? 'w-8 bg-purple-600' : 'w-2 bg-gray-600'
-                  }`}
+                  className={`h-1.5 rounded-full transition-all duration-300 ${activeSlide === index ? 'w-8 bg-purple-600' : 'w-2 bg-gray-600'
+                    }`}
                   aria-label={`Go to slide ${index + 1}`}
                 />
               ))}
