+++ utils/data/mock-data.ts
@@ -1,0 +1,228 @@
+import { Collection, ExpertCollection, NFT } from "../types/types";
+
+export const mockExpert: ExpertCollection[] = [
+    {
+        id: '1',
+        title: 'AstroFiction',
+        creator: {
+            name: 'Spaceone',
+            avatar: '/assets/Avatar.svg',
+        },
+        description: 'Lorem ipsum neque molestie nunc feugiat tincidunt vitae risus suscipit dictum tortor nunc sed.',
+        image: '/assets/space_girl.svg',
+        price: '1.63 ETH',
+        highestBid: '0.33 wETH',
+    },
+    {
+        id: '2',
+        title: 'Science Lab',
+        creator: {
+            name: 'Spaceone',
+            avatar: '/assets/Avatar.svg',
+        },
+        description: 'Lorem ipsum neque molestie nunc feugiat tincidunt vitae risus suscipit dictum tortor nunc sed.',
+        image: '/assets/science_lab_collection.png',
+        price: '2.45 ETH',
+        highestBid: '0.50 wETH',
+    },
+    {
+        id: '3',
+        title: 'Digital Future',
+        creator: {
+            name: 'NeoTech',
+            avatar: '/assets/Avatar.svg',
+        },
+        description: 'Exploring the boundaries of digital reality and blockchain technology in the modern era.',
+        image: '/assets/astronaut_collection.png', // Reusing for mock
+        price: '0.85 ETH',
+        highestBid: '0.12 wETH',
+    },
+    {
+        id: '4',
+        title: 'Quantum Art',
+        creator: {
+            name: 'Visionary',
+            avatar: '/assets/Avatar.svg',
+        },
+        description: 'A collection focusing on the intersection of quantum physics and visual aesthetics.',
+        image: '/assets/science_lab_collection.png', // Reusing for mock
+        price: '3.10 ETH',
+        highestBid: '0.75 wETH',
+    }
+];
+
+
+
+export const mockCollections: Collection[] = [
+    {
+        id: 1,
+        name: 'Spaceone',
+        items: 4,
+        price: '1.63',
+        image: '/nft3.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 2,
+        name: 'Spaceone',
+        items: 3,
+        price: '1.63',
+        image: '/nft1.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 3,
+        name: 'Spaceone',
+        items: 5,
+        price: '1.63',
+        image: '/nft2.svg',
+        avatar: '/nftAvatar3.png',
+    },
+];
+
+export const mocknftData: NFT[] = [
+    {
+        id: 1,
+        title: 'Science Lab',
+        timeLeft: '23+hrs',
+        rating: 4.7,
+        reviews: 4.7,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '1.63',
+        image: '/nft4.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 2,
+        title: 'Aoen.Net',
+        timeLeft: '23+hrs',
+        rating: 4.7,
+        reviews: 4.7,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '2.45',
+        image: '/nft7.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 3,
+        title: 'Flow Dymenisty',
+        timeLeft: '23+hrs',
+        rating: 4.7,
+        reviews: 4.7,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '0.89',
+        image: '/nft8.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 4,
+        title: 'Centry Networks',
+        timeLeft: '23+hrs',
+        rating: 4.7,
+        reviews: 4.7,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '3.21',
+        image: '/nft6.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 5,
+        title: 'Centry Networks',
+        timeLeft: '23+hrs',
+        rating: 4.8,
+        reviews: 4.8,
+        creator: 'Jane Anderson',
+        creatorRole: 'BlockchainExplorer',
+        price: '1.99',
+        image: '/nft8.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 6,
+        title: 'Flow Dymenisty',
+        timeLeft: '23+hrs',
+        rating: 4.6,
+        reviews: 4.6,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '1.35',
+        image: '/nft-lab.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 7,
+        title: 'Aoen.Net',
+        timeLeft: '23+hrs',
+        rating: 4.9,
+        reviews: 4.9,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '2.75',
+        image: '/nft6.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 8,
+        title: 'Science Lab',
+        timeLeft: '23+hrs',
+        rating: 4.5,
+        reviews: 4.5,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '0.75',
+        image: '/nft4.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 9,
+        title: 'Aoen.Net',
+        timeLeft: '23+hrs',
+        rating: 4.7,
+        reviews: 4.7,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '1.85',
+        image: '/nft6.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 10,
+        title: 'Flow Dymenisty',
+        timeLeft: '23+hrs',
+        rating: 4.8,
+        reviews: 4.8,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '3.50',
+        image: '/nft-neon.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 11,
+        title: 'Centry Networks',
+        timeLeft: '23+hrs',
+        rating: 4.6,
+        reviews: 4.6,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '2.10',
+        image: '/nft4.svg',
+        avatar: '/nftAvatar3.png',
+    },
+    {
+        id: 12,
+        title: 'Science Lab',
+        timeLeft: '23+hrs',
+        rating: 4.9,
+        reviews: 4.9,
+        creator: 'Spaceone',
+        creatorRole: 'BlockchainExplorer',
+        price: '1.20',
+        image: '/nft4.svg',
+        avatar: '/nftAvatar3.png',
+    },
+];
